Split config
