+++ src/cli/src/config/mod.rs
@@ -1,0 +1,2 @@
+pub mod setup;
+pub mod connection;
